+++ app/javascript/reactJS/discuss/comments/index.js
@@ -1,0 +1,1 @@
+import './comments_render'
